+++ model/schema_flag.go
@@ -41,13 +41,13 @@ func (s *Schema[T]) SetFlagsAt(rq *my.Request, condition qb.DualCondition[T], fi
 }
 
 // SetTxFlags updates the booleanField = flag as part of transaction at schema table affecting multiple rows
-func (s *Schema[T]) SetTxFlags(rq *my.Request, condition qb.DualCondition[T], field *bool, flag bool, numItems int) error {
-	return s.setFlagsAt(rq, condition, field, flag, numItems, s.Table, true)
+func (s *Schema[T]) SetTxFlags(rqtx *my.Request, condition qb.DualCondition[T], field *bool, flag bool, numItems int) error {
+	return s.setFlagsAt(rqtx, condition, field, flag, numItems, s.Table, true)
 }
 
 // SetTxFlagsAt updates the booleanField = flag as part of transaction at given table affecting multiple rows
-func (s *Schema[T]) SetTxFlagsAt(rq *my.Request, condition qb.DualCondition[T], field *bool, flag bool, numItems int, table string) error {
-	return s.setFlagsAt(rq, condition, field, flag, numItems, table, true)
+func (s *Schema[T]) SetTxFlagsAt(rqtx *my.Request, condition qb.DualCondition[T], field *bool, flag bool, numItems int, table string) error {
+	return s.setFlagsAt(rqtx, condition, field, flag, numItems, table, true)
 }
 
 // Common: create and execute UpdateQuery, which sets booleanField = flag at given table