+++ model/schema_insert_test.go
@@ -1,0 +1,17 @@
+package model
+
+import "testing"
+
+func TestInsert(t *testing.T) {
+	// TODO: Schema.Insert
+	// TODO: Schema.InsertAt
+	// TODO: Schema.InsertTx
+	// TODO: Schema.InsertTxAt
+}
+
+func TestInsertRows(t *testing.T) {
+	// TODO: Schema.InsertRows
+	// TODO: Schema.InsertRowsAt
+	// TODO: Schema.InsertTxRows
+	// TODO: Schema.InsertTxRowsAt
+}