+++ model/schema_update_test.go
@@ -1,0 +1,17 @@
+package model
+
+import "testing"
+
+func TestUpdateFields(t *testing.T) {
+	// TODO: Schema.UpdateFields
+	// TODO: Schema.UpdateFieldsAt
+	// TODO: Schema.UpdateTxFields
+	// TODO: Schema.UpdateTxFieldsAt
+}
+
+func TestUpdate(t *testing.T) {
+	// TODO: Schema.Update
+	// TODO: Schema.UpdateAt
+	// TODO: Schema.UpdateTx
+	// TODO: Schema.UpdateTxAt
+}